Added tests to Integration tests
