+++ tests/integration/data_access/test_entity_queries.py
@@ -3,6 +3,10 @@ import pytest
 
 from application.db.models import EntityOrm
 from application.data_access.entity_queries import get_entity_search
+from sqlalchemy.orm import Query
+from application.data_access.entity_queries import (
+    _apply_location_filters,
+)
 
 # set up logging
 logging.basicConfig(level=logging.DEBUG)
@@ -91,3 +95,87 @@ def test_get_entity_search_geometry_reference_queries_returns_correct_results(
 
     for entity in expected_entities:
         assert entity in [entity.entity for entity in results["entities"]], results
+
+
+def test__apply_location_filters_with_only_frz(db_session):
+    query = Query(EntityOrm)
+    result = _apply_location_filters(
+        db_session,
+        query,
+        params={
+            "longitude": "-0.2",
+            "latitude": "53.38",
+            "dataset": ["flood-risk-zone"],
+        },
+    )
+    sql_str = str(result.statement.compile(compile_kwargs={"literal_binds": True}))
+    assert "entity_subdivided" in sql_str
+
+
+def test__apply_location_filters_with_frz_and_others(db_session):
+    query = Query(EntityOrm)
+    result = _apply_location_filters(
+        db_session,
+        query,
+        params={
+            "longitude": "-0.2",
+            "latitude": "53.38",
+            "dataset": ["flood-risk-zone", "conservation-area"],
+        },
+    )
+    sql_str = str(result.statement.compile(compile_kwargs={"literal_binds": True}))
+    assert "entity_subdivided" not in sql_str
+
+
+def test__apply_location_filters_without_frz(db_session):
+    query = Query(EntityOrm)
+    result = _apply_location_filters(
+        db_session,
+        query,
+        params={
+            "longitude": "-0.2",
+            "latitude": "53.38",
+            "dataset": ["conservation-area"],
+        },
+    )
+    sql_str = str(result.statement.compile(compile_kwargs={"literal_binds": True}))
+    assert "entity_subdivided" not in sql_str
+
+
+def test__apply_location_filters_with_only_frz_geom(db_session):
+    query = Query(EntityOrm)
+    result = _apply_location_filters(
+        db_session,
+        query,
+        params={"geometry": "MULTIPOLYGON(-2.1 51.3)", "dataset": ["flood-risk-zone"]},
+    )
+    sql_str = str(result.statement.compile(compile_kwargs={"literal_binds": True}))
+    assert "entity_subdivided" in sql_str
+
+
+def test__apply_location_filters_with_frz_and_others_geom(db_session):
+    query = Query(EntityOrm)
+    result = _apply_location_filters(
+        db_session,
+        query,
+        params={
+            "geometry": "MULTIPOLYGON(-2.1 51.3)",
+            "dataset": ["flood-risk-zone", "conservation-area"],
+        },
+    )
+    sql_str = str(result.statement.compile(compile_kwargs={"literal_binds": True}))
+    assert "entity_subdivided" not in sql_str
+
+
+def test__apply_location_filters_without_frz_geom(db_session):
+    query = Query(EntityOrm)
+    result = _apply_location_filters(
+        db_session,
+        query,
+        params={
+            "geometry": "MULTIPOLYGON(-2.1 51.3)",
+            "dataset": ["conservation-area"],
+        },
+    )
+    sql_str = str(result.statement.compile(compile_kwargs={"literal_binds": True}))
+    assert "entity_subdivided" not in sql_str
